[WIP] Implementing TryFrom for array::ArraySchema
The idea here is that the "real" array scheams are defined in the array
module. The storage::ArraySchema is responsible for reading and writing
these to disk. Eventually I'll add a builder module that will be used by
the public APIs to create array schemas with the current API
definitions. I'll also likely have a capnp serializer to if and when I
get that far. So basically, three or so path ways in and out of the
array::ArraySchema which contains all of the various API bits required
by the rest of our core logic.