[WIP] Adding ArrayDirectory equivalent
I realized I needed to know how to find fragments in all of the old
arrays so I started writing an equivalent to the ArrayDirectory to map
the logic. Turns out its even more gnarly than I thought so its going to
be a bit of a detour.

I think for the most part I'm going to ignore updates and deletes for
the time being and just focus on the happy paths so that I can start
working through the fragment metadadata deserialization.